+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add queued tasks end timeout support to worker node termination.
+
 ## [3.1.4] - 2023-12-18
 
 ### Fixed
@@ -23,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Wait for queued tasks to end at worker termination.
+- Wait for queued tasks to end at worker node termination.
 
 ## [3.1.1] - 2023-12-16